Improve test reliability and fix workstation domain trust issue
