+++ .github/workflows/release.yaml
@@ -43,10 +43,21 @@ jobs:
         run: dotnet --info
 
       - name: Pack
-        run: dotnet pack -c release /p:ContinuousIntegrationBuild=true
+        run: |
+          dotnet pack -c release /p:ContinuousIntegrationBuild=true
+          cp -r nupkg nupkg_pack
+          dotnet clean
 
       - name: Build (AOT)
-        run: dotnet publish -c release /p:UseAot=1 /p:ContinuousIntegrationBuild=true
+        run: |
+          dotnet publish -c release /p:UseAot=1 /p:ContinuousIntegrationBuild=true
+          cp -r bin bin_aot
+          dotnet clean
+
+      - name: Build (release)
+        run: |
+          dotnet publish -c release /p:ContinuousIntegrationBuild=true
+          cp -r nupkg nupkg_release
 
       - name: Create release
         if: ${{ github.ref == 'refs/heads/main' && startsWith(github.event.head_commit.message, 'chore(release)') }}
@@ -54,10 +65,10 @@ jobs:
         with:
           draft: true
           name: "${{ steps.get_version.outputs.version }}"
-          files: ForceOps/bin/Release/net8.0/win-x64/publish/ForceOps.exe
+          files: ForceOps/bin_aot/Release/net8.0/win-x64/publish/ForceOps.exe
           tag_name: "${{ steps.get_version.outputs.version }}"
 
       - name: Publish NuGet
         if: ${{ github.ref == 'refs/heads/main' && startsWith(github.event.head_commit.message, 'chore(release)') }}
         run: |
-          dotnet nuget push -k ${{ secrets.NUGET_AUTH_TOKEN }} -s https://api.nuget.org/v3/index.json ForceOps/nupkg/ForceOps.${{ steps.get_version.outputs.version }}.nupkg ForceOps.Lib/nupkg/ForceOps.Lib.${{ steps.get_version.outputs.version }}.nupkg --skip-duplicate
+          dotnet nuget push -k ${{ secrets.NUGET_AUTH_TOKEN }} -s https://api.nuget.org/v3/index.json ForceOps/nupkg_pack/ForceOps.${{ steps.get_version.outputs.version }}.nupkg ForceOps.Lib/nupkg_release/ForceOps.Lib.${{ steps.get_version.outputs.version }}.nupkg --skip-duplicate
