+++ ForceOps.Lib/ForceOps.Lib.csproj
@@ -11,7 +11,7 @@
 		<None Include="..\LICENSE.txt" Pack="true" PackagePath="" />
 	</ItemGroup>
 	<ItemGroup>
-		<PackageReference Include="LockChecker" Version="0.9.30-g4f01a64a51" />
+		<PackageReference Include="LockChecker" Version="0.9.33-gf4fccad662" />
 		<PackageReference Include="Serilog" Version="3.0.1" />
 		<PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
 	</ItemGroup>
